refactor: update contract creation process and UI components
- Refactored the contract creation logic to accept a name parameter alongside content.
- Enhanced the New Contract page UI with improved state management and user feedback.
- Removed the deprecated start-drafting page and updated routing to the new contract creation page.
- Adjusted JSON templates for better alignment with new contract requirements.
- Improved file upload handling and user experience in the contract drafting process.
